+++ Music-Bot-Telegram.Data/Models/ActionType.cs
@@ -4,7 +4,7 @@ namespace Music_Bot_Telegram.Data.Models;
 
 public class ActionType : IEntity
 {
-    [Key] public int Id { get; set; }
+    [Key] public long Id { get; set; }
     /* unique */ public string Name { get; set; }
     
     public ActionType()
