+++ Music-Bot-Telegram.Data/Models/User.cs
@@ -4,7 +4,7 @@ namespace Music_Bot_Telegram.Data.Models;
 
 public class User : IEntity
 {
-    [Key] public int Id { get; set; }
+    [Key] public long Id { get; set; }
     public bool IsAdmin { get; set; } = false;
     public bool IsBanned { get; set; } = false;
     public bool IsActiveSession { get; set; } = false;
